Follow the same semantics as webapp for "plugins"
Summary:
In the initial commit, I added the 'require' statements at the end
of the module being required instead the file doing the requiring.
This diff changes the insertion point to be the end of the file doing
the requiring.  This matches the same semantics as webapp.

Test Plan: npm test

Reviewers: john jaqueline

Subscribers: csilvers jeff mattdunnrankin